Trait authzen_diesel::prelude::DbEntity

pub trait DbEntity: Sized + Send + 'static {
    type Raw: HasTable<Table = Self::Table> + TryInto<Self> + Send + 'static;
    type Table: Table + QueryId + Send;
    type Id: AsExpression<<<Self::Table as Table>::PrimaryKey as Expression>::SqlType>
       where <<Self::Table as Table>::PrimaryKey as Expression>::SqlType: SqlType;
}

Base trait which unites common associated types used by the common sql operations in diesel (i.e. select, insert, update and delete).

Required Associated Types

type Raw: HasTable<Table = Self::Table> + TryInto<Self> + Send + 'static

An equivalent representation of this entity which has diesel trait implementations.

type Table: Table + QueryId + Send

The table this entity represents.

type Id: AsExpression<<<Self::Table as Table>::PrimaryKey as Expression>::SqlType> where <<Self::Table as Table>::PrimaryKey as Expression>::SqlType: SqlType

The type of this entity’s table’s primary key note that this type should be equivalent diesel sql_type representation of the type of the primary key field in Self::Raw.